# 8302623: jarsigner - use BufferedOutputStream to improve performance while creating the signed jar